src/app/core/services/paginator-intl.service.ts
import { Injectable } from '@angular/core';
import { MatPaginatorIntl } from '@angular/material/paginator';

@Injectable({
  providedIn: 'root'
})
export class PaginatorIntlService extends MatPaginatorIntl {
  firstPageLabel = 'Primeira página';
  itemsPerPageLabel = 'Exibir';
  lastPageLabel = 'Última página';
  nextPageLabel = 'Próxima página';
  previousPageLabel = 'Página anterior';

  getRangeLabel = (page: number, pageSize: number, length: number) => {
    if (length === 0 || pageSize === 0) {
      return `0 de ${length}`;
    }
    length = Math.max(length, 0);
    const endPage = Math.ceil(length / pageSize);
    return `Página ${page + 1} de ${endPage}`;
  };
}
